Reorganise directories to clean up project root.
