+++ CodeCaster.SerializeThis/Package/[Content_Types].xml
