+++ CodeCaster.SerializeThis/Package/source.extension.vsixmanifest
